[Decomp][Achievement] Cohesion split: extract GlobalMgr + criteria-requirement validation to Achievement*.cpp
Move two cohesive, self-contained classes out of the 3,304-line
AchievementMgr.cpp into dedicated TUs (verbatim move, no behaviour change;
AchievementMgr.h unchanged, CMake file(GLOB WorldHandlers/*.cpp) auto-registers
them):

  AchievementGlobalMgr.cpp            the AchievementGlobalMgr registry/loaders
                                      (LoadAchievementCriteriaList/ReferenceList/
                                      CriteriaRequirements/CompletedAchievements/
                                      Rewards/RewardLocales + accessors)
  AchievementCriteriaRequirement.cpp  AchievementCriteriaRequirement[Set]
                                      validation (IsValid + Meets gating)

AchievementMgr.cpp 3,304 -> 2,382. The per-player AchievementMgr class (incl.
the large UpdateAchievementCriteria dispatcher and its file-local lookup tables)
stays for a focused follow-up.
